+++ shipcat_cli/tests/webhooks.rs
@@ -8,14 +8,14 @@ use shipcat;
 use shipcat_definitions;
 
 use crate::shipcat::webhooks;
-use crate::shipcat_definitions::{Config, ConfigType};
+use crate::shipcat_definitions::{Config, ConfigState};
 
 #[test]
 // #[serial] not available still :( so have to concatenate two tests messing with evars
 fn webhooks_ensure_requirements() {
     common::setup();
 
-    let (_conf, reg) = Config::new(ConfigType::Filtered, "dev-uk").unwrap();
+    let (_conf, reg) = Config::new(ConfigState::Filtered, "dev-uk").unwrap();
 
     env::set_var("SHIPCAT_AUDIT_REVISION", "egrevision");
     env::set_var("SHIPCAT_AUDIT_CONTEXT_ID", "egcontextid");